Implemented priority matrix method which return a list of suggested tools ids

+++ backend/src/main/java/com/dsec/backend/repository/ToolRepository.java
@@ -3,5 +3,5 @@ package com.dsec.backend.repository;
 import com.dsec.backend.entity.ToolEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToolRepository  extends JpaRepository<ToolEntity, Long> {
+public interface ToolRepository  extends JpaRepository<ToolEntity, Integer> {
 }
